Module higher_effect::io

Structs

IO

An IO monad.

Functions

append_file

Append to an already existing file on the filesystem.

get_contents

Read the entire contents of the standard input.

get_line

Read a line from the console.

print

Print any value which implements Display to the console, followed by a newline.

put_str

Print a string to the console.

put_str_ln

Print a string and a newline to the console.

read_file

Read a file from the filesystem.

write_file

Write a file to the filesystem.