Struct higher_effect::Effect

pub struct Effect<'a, A> { /* private fields */ }

An effect monad.

This is essentially just a newtype for a boxed Future that implements Monad, so that you can treat your Futures like they’re Monads with the run! macro and all the category theory you like.

To turn a Future into an Effect monad, use From and Into:

let my_future = async { "Hello Joe!" };
let my_effect = Effect::from(my_future);

Effects can be awaited like they’re futures, because they implement IntoFuture:

assert_eq!(my_effect.await, "Hello Joe!");

You can compose effects using the run! macro:

run! {
    // Lift the value 1 into the Effect monad
    x <= Effect::pure(1);
    // Create an effect from an async block returning the value 2
    y <= Effect::from(async { 2 });
    // Perform a computation in an async block using the previously bound values
    z <= Effect::from(async move { x + y });
    // Compute the result and await it
    yield x + y + z
}.await

Implementations

impl<'a, A> Effect<'a, A>

pub fn run(self) -> A

Run the effect to completion, blocking the current thread.

pub fn ready(value: A) -> Selfwhere
A: 'a,

Construct an effect that resolves immediately to the given value.

pub fn select<B>(
eff1: Effect<'a, A>,
eff2: Effect<'a, B>
) -> Effect<'a, Either<(A, Effect<'a, B>), (B, Effect<'a, A>)>>where
A: 'a,
B: 'a,

Runs two effects in parallel, returning the result of whichever finishes first.

pub fn select_all<I>(iter: I) -> Effect<'a, (A, usize, Vec<Effect<'a, A>>)>where
A: 'a,
I: IntoIterator<Item = Effect<'a, A>> + 'a,

Runs a list of effects in parallel, returning the result of whichever finishes first.

pub fn join<B>(eff1: Effect<'a, A>, eff2: Effect<'a, B>) -> Effect<'a, (A, B)>where
A: 'a,
B: 'a,

Runs two effects in parallel, returning the results of both.

pub fn join3<B, C>(
eff1: Effect<'a, A>,
eff2: Effect<'a, B>,
eff3: Effect<'a, C>
) -> Effect<'a, (A, B, C)>where
A: 'a,
B: 'a,
C: 'a,

Runs three effects in parallel, returning the results of all three.

pub fn join4<B, C, D>(
eff1: Effect<'a, A>,
eff2: Effect<'a, B>,
eff3: Effect<'a, C>,
eff4: Effect<'a, D>
) -> Effect<'a, (A, B, C, D)>where
A: 'a,
B: 'a,
C: 'a,
D: 'a,

Runs four effects in parallel, returning the results of all four.

pub fn join_all<I>(iter: I) -> Effect<'a, Vec<A>>where
I: IntoIterator<Item = Effect<'a, A>> + 'a,

Runs a list of effects in parallel, returning a list of their results when they have all resolved.