Struct higher_effect::IO

pub struct IO<'a, A, E> { /* private fields */ }

An IO monad.

This wraps a Future similarly to the Effect monad, but the wrapped effect must return a Result. Just as with Result’s Bind implementation, this also short circuits a chain of computations if a step produces an error, resolving immediately to that error.

You can construct an IO monad from a Future which returns a Result:

let my_io_monad = IO::<&str, &str>::from(async { Ok("Hello Joe!") });

You can .await an IO monad and get a Result back:

assert_eq!(my_io_monad.await, Ok("Hello Joe!"));

You can run your IO monads on a thread local executor using the run() method. Naturally, we also provide a version of Haskell’s putStrLn so that we can implement the canonical hello world in monadic Rust:

put_str_ln("Hello Simon!").run();

Because IO implements Bind, you can chain async IO operations together using the run! macro:

run! {
    put_str("What is your name? ");
    name <= get_line();
    put_str_ln(format!("Hello {}!", name))
}.run();

Implementations

impl<'a, A, E> IO<'a, A, E>

pub fn run(self) -> Result<A, E>

Run the effect to completion, blocking the current thread.

pub fn map_ok<B, F>(self, f: F) -> IO<'a, B, E>where
A: 'a,
B: 'a,
E: 'a,
F: FnOnce(A) -> B + 'a,

Transform the effect’s success value using the given function.

If the effect fails, the function is not applied.

This is identical to Functor::fmap, except it accepts a FnOnce instead of a Fn.

pub fn map_error<B, F>(self, f: F) -> IO<'a, A, B>where
A: 'a,
B: 'a,
E: 'a,
F: FnOnce(E) -> B + 'a,

Transform the effect’s error value using the given function.

If the effect succeeds, the function is not applied.

This is identical to Bifunctor::rmap, except it accepts a FnOnce instead of a Fn.

pub fn is_err(&self) -> bool

Test whether this effect is known to have already failed.

pub fn try_select<B, EB>(
eff1: IO<'a, A, E>,
eff2: IO<'a, B, EB>
) -> IO<'a, Either<(A, IO<'a, B, EB>), (B, IO<'a, A, E>)>, Either<(E, IO<'a, B, EB>), (EB, IO<'a, A, E>)>>where
A: 'a,
B: 'a,
E: 'a,
EB: 'a,

Run two effects in parallel, returning the result of whichever finishes first, or the error from whichever fails first.

pub fn select_ok<I>(iter: I) -> IO<'a, (A, Vec<IO<'a, A, E>>), E>where
A: 'a,
E: 'a,
I: IntoIterator<Item = IO<'a, A, E>> + 'a,

Run a list of effects in parallel, returning the result of whichever finishes first, or the first error to occur.

pub fn try_join<B>(eff1: IO<'a, A, E>, eff2: IO<'a, B, E>) -> IO<'a, (A, B), E>where
A: 'a,
B: 'a,
E: 'a,

Run two effects in parallel, returning the results of both, or the error from whichever fails first.

pub fn try_join3<B, C>(
eff1: IO<'a, A, E>,
eff2: IO<'a, B, E>,
eff3: IO<'a, C, E>
) -> IO<'a, (A, B, C), E>where
A: 'a,
B: 'a,
C: 'a,
E: 'a,

Run three effects in parallel, returning the results of all three, or the error from whichever fails first.

pub fn try_join4<B, C, D>(
eff1: IO<'a, A, E>,
eff2: IO<'a, B, E>,
eff3: IO<'a, C, E>,
eff4: IO<'a, D, E>
) -> IO<'a, (A, B, C, D), E>where
A: 'a,
B: 'a,
C: 'a,
D: 'a,
E: 'a,

Run four effects in parallel, returning the results of all four, or the error from whichever fails first.

pub fn try_join_all<I>(iter: I) -> IO<'a, Vec<A>, E>where
I: IntoIterator<Item = IO<'a, A, E>> + 'a,

Run a list of effects in parallel, returning a list of their results, or the error from whichever fails first.